Fall back to check_cxx_source_compiles when cross-compiling
On certain systems, code using std::fs can appear to link successfully
but fail at runtime. But always checking the runtime result of the check
executable precludes cross-compilation. So now when cross-compiling, we
only check that the test executable compiles - otherwise, we try to run
the resulting executable as well.

Fixes #10